Bug fix. Responsiveness: Button aspect-ratio
Buttons without labels are supposed to maintain a 1/1
aspect-ratio at all times, but instead they were "squishing up"
on narrow screens (most noticable in the colour-palette stories,
where the buttons are displayed in a table).

Button squishing was caused by the SVG icon shrinking to fit,
which in turn was caused by the max-width: 100% applied to all
image tags in reset.css.

Overriding the max-width fixed the issue.