Bug fix. Responsive & Aria: SelectTags and Search
Fixed Bug #1: SelectTags
Grid blow-out. On a small enough screen, the clear and toggle
buttons were "pushed" outside of the visible area.

Fixed Bug #2: SelectTags and Search
Added aria properties to buttons, so it's possible to tell
the difference when you're a screenreader (or, indeed, a dev
looking in the inspector, trying to fix a grid blowout).

Small change: SelectTags and Search
Writing the aria-labels made me realise "clear" better describes
that button's purpose than "close", so did some renaming.