BugFix. FileUploader responsive & accessible.
Bug #1:
Made the contents of the FileUploader responsive, instead of just
the outsides. (Previously: the outside was shrinking to fit nicely,
but the content was overflowing and disappearing.)

Consequence:
This broke the solution for the "drag-and-drop flickering" bug.
New solution = now the "drag starting" element records its height
in state when a drag begins, so this can be passed to (and matched by)
the "drag ending" element.

Bug #2:
Enabled use of the FileUploader while using a screenreader and/or a
keyboard.